Add titles to each page and fix buttons
