+++ components/ui/buttons/all.tsx
@@ -26,28 +26,21 @@ export function JoinDiscordButton() {
   )
 }
 
-type CustomProps = {
-  name: string;
-  link: string;
-  ring: string;
-  border: string;
-  bg: string;
-  hoverBg: string;
-}
-
-export function CustomButton({ name, link, ring, border, bg, hoverBg }: CustomProps) {
-  return (<>
-    <Container tapScale={0.95}>
-      <Link href={link} className={`ring-[${ring}] btn border-[${border}] bg-[${bg}] hover:bg-[${hoverBg}] text-white font-semibold px-6 px-5.5 py-2.5 shadow-none text-sm`}>{name}</Link>
-    </Container>
-  </>)
-}
-
-export function ExploreScriptsButton({}) {
+export function ExploreScriptsButton({ }) {
   return (<>
     <Container tapScale={0.95}>
       <Link href={"/scripts"} className="ring-[#21582d] btn border-[#72da89] bg-[#247a37] hover:bg-[#1b5828] text-white font-semibold px-6 py-2.5 text-sm group">
         Explore scripts
+      </Link>
+    </Container>
+  </>)
+}
+
+export function MoreScreenshotsButton({ }) {
+  return (<>
+    <Container tapScale={0.95}>
+      <Link href={"/gallery"} className="ring-[#4f2158] btn border-[#7d3868] bg-[#7d368b] hover:bg-[#511c5c] text-white font-semibold px-6 py-2.5 text-sm group">
+        More screenshots
       </Link>
     </Container>
   </>)
